requirements.txt
future>=1.0.0
openai
tiktoken
networkx
graspologic
nano-vectordb
hnswlib
xxhash
tenacity
dspy-ai
neo4j
aioboto3
google-genai
gradio
ipykernel
milvus-lite
pymilvus
ollama
python-dotenv
transformers
sentence-transformers
pyvi